+++ src/phosphorus/construction/sdist.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import tarfile
-from pathlib import Path
 from typing import TYPE_CHECKING
 
 from phosphorus.construction.base import Builder
@@ -10,7 +9,8 @@ from phosphorus.lib.licenses import get_license_files
 from phosphorus.lib.zipped_file import ArchiveFile
 
 if TYPE_CHECKING:
-    from collections.abc import Iterator
+    from collections.abc import Iterable, Iterator
+    from pathlib import Path
 
 
 class SdistBuilder(Builder):
@@ -20,23 +20,34 @@ class SdistBuilder(Builder):
     def filename(self) -> str:
         return f"{self.base_name}.tar.gz"
 
-    def collect_files(self, temp_dir: Path) -> list[ArchiveFile]:
-        output: list[ArchiveFile] = []
+    def package_files(self, _temp_dir: Path) -> Iterator[ArchiveFile]:
         base_dir = self.meta.base_dir
         for package in self.meta.package_paths:
-            output.extend(
-                ArchiveFile.from_file(source=file, target=file.relative_to(base_dir))
-                for file in package.path.glob("**/*")
-                if file.is_file()
-            )
-        output.extend(
-            ArchiveFile.from_file(source=source, target=target)
-            for source, target in self.non_package_files(temp_dir)
+            for file in package.absolute_path.rglob("*"):
+                if file.is_file():
+                    yield ArchiveFile.from_file(source=file, base_dir=base_dir)
+
+    def non_package_files(self, _temp_dir: Path) -> Iterator[ArchiveFile]:
+        base_dir = self.meta.base_dir
+        yield ArchiveFile.from_file(
+            base_dir.joinpath(pyproject_base_name), base_dir=base_dir
         )
-        return output
+
+        for license_file in get_license_files(base_dir):
+            yield ArchiveFile.from_file(license_file, base_dir=base_dir)
+
+        if self.meta.readme.read_text():
+            yield ArchiveFile.from_file(self.meta.readme, base_dir=base_dir)
+
+    def get_info_file(self, temp_dir: Path, _data: str = "") -> ArchiveFile:
+        pkg_info = temp_dir.joinpath("PKG-INFO")
+        with pkg_info.open("w") as file:
+            file.writelines(f"{line}\n" for line in self.get_metadata_content())
+
+        return ArchiveFile.from_file(source=pkg_info, base_dir=temp_dir)
 
     def write_files(
-        self, files: list[ArchiveFile], package: Path, _temp_dir: Path
+        self, files: Iterable[ArchiveFile], package: Path, temp_dir: Path
     ) -> None:
         with tarfile.open(package, "w:gz") as tar:
             for archive_file in sorted(files):
@@ -44,19 +55,5 @@ class SdistBuilder(Builder):
                     archive_file.tar_info, archive_file.absolute_path.open("rb")
                 )
 
-    def non_package_files(self, temp_dir: Path) -> Iterator[tuple[Path, Path]]:
-        base_dir = self.meta.base_dir
-
-        yield base_dir.joinpath(pyproject_base_name), Path(pyproject_base_name)
-
-        for license_file in get_license_files(base_dir):
-            yield license_file, license_file.relative_to(base_dir)
-
-        if self.meta.readme.read_text():
-            yield self.meta.readme, self.meta.readme.relative_to(base_dir)
-
-        pkg_info = temp_dir.joinpath("PKG-INFO")
-        with pkg_info.open("w") as file:
-            file.writelines(f"{line}\n" for line in self.get_metadata_content())
-
-        yield pkg_info, pkg_info.relative_to(temp_dir)
+            info_file = self.get_info_file(temp_dir)
+            tar.addfile(info_file.tar_info, info_file.absolute_path.open("rb"))